+++ baseLevelable.go
@@ -12,9 +12,9 @@ type BaseLevelable struct {
 }
 
 // GetPrice returns the price to build the given level
-func (b BaseLevelable) GetPrice(level int) Resources {
-	tmp := func(baseCost int, increaseFactor float64, level int) int {
-		return int(float64(baseCost) * math.Pow(increaseFactor, float64(level-1)))
+func (b BaseLevelable) GetPrice(level int64) Resources {
+	tmp := func(baseCost int64, increaseFactor float64, level int64) int64 {
+		return int64(float64(baseCost) * math.Pow(increaseFactor, float64(level-1)))
 	}
 	return Resources{
 		Metal:     tmp(b.BaseCost.Metal, b.IncreaseFactor, level),
