+++ baseTechnology.go
@@ -11,8 +11,8 @@ type BaseTechnology struct {
 }
 
 // ConstructionTime returns the duration it takes to build given technology
-func (b BaseTechnology) ConstructionTime(level, universeSpeed int, facilities Facilities) time.Duration {
-	price := b.GetPrice(level)
+func (b BaseTechnology) ConstructionTime(level, universeSpeed int64, facilities Facilities) time.Duration {
+	price := b.GetPrice(int64(level))
 	metalCost := float64(price.Metal)
 	crystalCost := float64(price.Crystal)
 	researchLabLvl := float64(facilities.ResearchLab)
@@ -22,6 +22,6 @@ func (b BaseTechnology) ConstructionTime(level, universeSpeed int, facilities Fa
 }
 
 // GetLevel returns current level of a technology
-func (b BaseTechnology) GetLevel(_ LazyResourcesBuildings, _ LazyFacilities, lazyResearches LazyResearches) int {
+func (b BaseTechnology) GetLevel(_ LazyResourcesBuildings, _ LazyFacilities, lazyResearches LazyResearches) int64 {
 	return lazyResearches().ByID(b.ID)
 }
