+++ cruiser.go
@@ -14,10 +14,10 @@ func newCruiser() *cruiser {
 	s.BaseCargoCapacity = 800
 	s.BaseSpeed = 15000
 	s.FuelConsumption = 300
-	s.RapidfireFrom = map[ID]int{BattlecruiserID: 4, DeathstarID: 33}
-	s.RapidfireAgainst = map[ID]int{EspionageProbeID: 5, SolarSatelliteID: 5,
+	s.RapidfireFrom = map[ID]int64{BattlecruiserID: 4, DeathstarID: 33}
+	s.RapidfireAgainst = map[ID]int64{EspionageProbeID: 5, SolarSatelliteID: 5,
 		LightFighterID: 6, RocketLauncherID: 10}
 	s.Price = Resources{Metal: 20000, Crystal: 7000, Deuterium: 2000}
-	s.Requirements = map[ID]int{ShipyardID: 5, ImpulseDriveID: 4, IonTechnologyID: 2}
+	s.Requirements = map[ID]int64{ShipyardID: 5, ImpulseDriveID: 4, IonTechnologyID: 2}
 	return s
 }
