+++ hyperspaceDrive.go
@@ -10,6 +10,6 @@ func newHyperspaceDrive() *hyperspaceDrive {
 	b.ID = HyperspaceDriveID
 	b.IncreaseFactor = 2.0
 	b.BaseCost = Resources{Metal: 10000, Crystal: 20000, Deuterium: 6000}
-	b.Requirements = map[ID]int{HyperspaceTechnologyID: 3}
+	b.Requirements = map[ID]int64{HyperspaceTechnologyID: 3}
 	return b
 }
