+++ lightFighter.go
@@ -14,9 +14,9 @@ func newLightFighter() *lightFighter {
 	l.BaseCargoCapacity = 50
 	l.BaseSpeed = 12500
 	l.FuelConsumption = 20
-	l.RapidfireFrom = map[ID]int{CruiserID: 6, DeathstarID: 200}
-	l.RapidfireAgainst = map[ID]int{EspionageProbeID: 5, SolarSatelliteID: 5}
+	l.RapidfireFrom = map[ID]int64{CruiserID: 6, DeathstarID: 200}
+	l.RapidfireAgainst = map[ID]int64{EspionageProbeID: 5, SolarSatelliteID: 5}
 	l.Price = Resources{Metal: 3000, Crystal: 1000}
-	l.Requirements = map[ID]int{ShipyardID: 1, CombustionDriveID: 1}
+	l.Requirements = map[ID]int64{ShipyardID: 1, CombustionDriveID: 1}
 	return l
 }
