+++ lightLaser.go
@@ -12,7 +12,7 @@ func newLightLaser() *lightLaser {
 	d.StructuralIntegrity = 2000
 	d.ShieldPower = 25
 	d.WeaponPower = 100
-	d.RapidfireFrom = map[ID]int{DestroyerID: 10, BomberID: 20, DeathstarID: 200}
-	d.Requirements = map[ID]int{ShipyardID: 2, LaserTechnologyID: 3}
+	d.RapidfireFrom = map[ID]int64{DestroyerID: 10, BomberID: 20, DeathstarID: 200}
+	d.Requirements = map[ID]int64{ShipyardID: 2, LaserTechnologyID: 3}
 	return d
 }
