+++ missileSilo.go
@@ -10,6 +10,6 @@ func newMissileSilo() *missileSilo {
 	b.ID = MissileSiloID
 	b.IncreaseFactor = 2.0
 	b.BaseCost = Resources{Metal: 20000, Crystal: 20000, Deuterium: 1000}
-	b.Requirements = map[ID]int{ShipyardID: 1}
+	b.Requirements = map[ID]int64{ShipyardID: 1}
 	return b
 }
