+++ plasmaTurret.go
@@ -12,6 +12,6 @@ func newPlasmaTurret() *plasmaTurret {
 	d.StructuralIntegrity = 100000
 	d.ShieldPower = 300
 	d.WeaponPower = 3000
-	d.Requirements = map[ID]int{ShipyardID: 8, PlasmaTechnologyID: 7}
+	d.Requirements = map[ID]int64{ShipyardID: 8, PlasmaTechnologyID: 7}
 	return d
 }
