+++ weaponsTechnology.go
@@ -10,6 +10,6 @@ func newWeaponsTechnology() *weaponsTechnology {
 	b.ID = WeaponsTechnologyID
 	b.IncreaseFactor = 2.0
 	b.BaseCost = Resources{Metal: 800, Crystal: 200}
-	b.Requirements = map[ID]int{ResearchLabID: 4}
+	b.Requirements = map[ID]int64{ResearchLabID: 4}
 	return b
 }
